enable benchmark using libtest
