+++ Cargo.toml
@@ -31,3 +31,4 @@ default-features = false
 png-encoding = ["deflate"]
 default = ["png-encoding"]
 unstable = []
+benchmarks = []
